S10P22A309-273 Feat : 여행->은행 api object mapper 적용 및 수정

+++ BE/trip-together/src/main/java/com/ssafy/triptogether/auth/data/response/TokenInfo.java
@@ -10,6 +10,6 @@ public record TokenInfo(
     @JsonProperty("created_at")
     Long createdAt,
     @JsonProperty("expires_in")
-    Integer expiresIn
+    Long expiresIn
 ) {
 }
